fix points

+++ database/migrations/2014_10_12_000000_create_users_table.php
@@ -25,7 +25,7 @@ return new class extends Migration
             $table->foreignId('role_id');
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
-            $table->decimal('points')->nullable();
+            $table->integer('points')->nullable();
             $table->rememberToken();
             $table->timestamps();
         });
